# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Chat {
  id: number;
  title: string;
  model: string;
  totalTokens: number;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateChatRequest {
  title: string;
  model: string;
}

export interface UpdateChatRequest {
  title?: string;
  model?: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
  system: 'system',
} as const;

export interface Message {
  id: number;
  chatId: number;
  role: MessageRole;
  content: string;
  tokensUsed: number;
  createdAt: string;
}

export interface StreamRequest {
  content: string;
  images?: string[];
}

export interface FilePathRequest {
  path: string;
}

export type FileEntryType = typeof FileEntryType[keyof typeof FileEntryType];


export const FileEntryType = {
  file: 'file',
  directory: 'directory',
} as const;

export interface FileEntry {
  name: string;
  path: string;
  type: FileEntryType;
  size?: number | null;
}

export interface FileContent {
  content: string;
}

export interface WriteFileRequest {
  path: string;
  content: string;
}

export interface SuccessResponse {
  ok: boolean;
}

export interface ExecuteCommandRequest {
  command: string;
  cwd?: string;
}

export interface CommandResult {
  stdout: string;
  stderr: string;
  exitCode: number;
}

export interface ModelInfo {
  id: string;
  name: string;
}

export type GetSettings200 = {[key: string]: string};

export type SaveSettingsBody = {[key: string]: string};

export type SaveSettings200 = {[key: string]: string};